PaymentMethod controller
Renamed controller.

Added TODO info for team member.
